feat: se cambio el nombre del componente AsideBar a SideBar
