+++ frontend/src/components/sections/DashboardPatient/DashboardPatient.tsx
@@ -1,13 +1,13 @@
 import { Grid2 } from "@mui/material";
 import { DashboardPatientStyles } from "./DashboardPatientStyles";
-import { AsideBar } from "@ui/AsideBar/AsideBar";
+import { SideBar } from "@ui/SideBar/SideBar";
 import { PatientMainContent } from "./PatientMainContent/PatientMainContent";
 import { PATIENT_LINKS } from "./PATIENT_LINKS.tsx";
 
 export const DashboardPatient = () => {
     return (
         <Grid2 container sx={DashboardPatientStyles.container}>
-            <AsideBar links={PATIENT_LINKS} />
+            <SideBar links={PATIENT_LINKS} />
             <PatientMainContent />
         </Grid2>
     );
