+++ frontend/src/components/sections/Medico/HistorialMedico/HistorialMedico.tsx
@@ -3,7 +3,7 @@ import { MEDIC_LINKS } from "../MEDIC_LINKS";
 
 export const HistorialMedico = () => {
     return (
-        <SectionWrapper asideBarItems={MEDIC_LINKS}>
+        <SectionWrapper sideBarItems={MEDIC_LINKS}>
             <></>
         </SectionWrapper>
     );
