+++ frontend/src/ui/SideBar/NavigationItem/NavigationItemStyle.ts
